feat(components): utilized utils and modified fetcher
