+++ src/hooks/index.js
@@ -6,7 +6,7 @@ import { useQuery } from "@tanstack/react-query";
 export const useTrendings = () =>
   useQuery({
     queryKey: ["trendings"],
-    queryFn: async () => await fetchTrending("tv"),
+    queryFn: async () => await fetchTrending(),
   });
 
 export const useGenres = (mediaType = "movie") =>
